+++ src/router.tsx
@@ -1,11 +1,13 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import MinglePage from "./Main/mingle";
+import LoginPage from "./pages/Login/login";
+import MinglePage from "./pages/Main/mingle";
 
 export default function Router() {
     return (
         <BrowserRouter>
             <Routes>
                 <Route path="/" element={<MinglePage />} />
+                <Route path="/login" element={<LoginPage />} />
             </Routes>
         </BrowserRouter>
     );
